create util library with json function
