+++ lib/asf-stac-util/setup.py
@@ -1,0 +1,10 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='asf-stac-util',
+    license='BSD',
+    include_package_data=True,
+    install_reqires=[],
+    python_requires='~=3.9',
+    packages=find_packages(),
+)
