+++ requirements.txt
@@ -1,5 +1,6 @@
 -r requirements-apps-api.txt
 -r requirements-run-codebuild.txt
+./lib/asf-stac-util/
 boto3==1.26.21
 cfn-lint==0.72.1
 flake8==6.0.0
