+++ tests/test_create_coherence_items.py
@@ -24,26 +24,6 @@ def test_season_datetime_averages():
            == timedelta(days=45)
 
 
-def test_jsonify_stac_item():
-    assert create_coherence_items.jsonify_stac_item(
-        {
-            'str_field': 'foo',
-            'int_field': 5,
-            'float_field': 3.1,
-            'bool_field': True,
-            'null_field': None,
-            'dict_field': {'str_field': 'bar'},
-            'list_field': [[1, 2], [3, 4]],
-            'tuple_field': ((1, 2), (3, 4)),
-            'datetime_field': datetime(2022, 11, 30, 12, tzinfo=timezone.utc),
-        }
-    ) == (
-        '{"str_field": "foo", "int_field": 5, "float_field": 3.1, "bool_field": true, "null_field": null, '
-        '"dict_field": {"str_field": "bar"}, "list_field": [[1, 2], [3, 4]], "tuple_field": [[1, 2], [3, 4]], '
-        '"datetime_field": "2022-11-30T12:00:00Z"}'
-    )
-
-
 def test_create_stac_item_N00E005_124D_inc():
     assert create_coherence_items.create_stac_item('data/tiles/N00E005/N00E005_124D_inc.tif', 'foo.com/') \
            == {
